requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.8.0
python-dotenv>=1.0.1
huggingface_hub>=0.24.0
pinecone>=5.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
pypdf>=4.2.0
sentence-transformers>=3.0.1
rank-bm25>=0.2.2
torch>=2.3.0
